Fetch commits using Authorization header
`nix flake info` calls a github API that requires authorization when the
repository in question is private. Currently this request fails with a
404.

This commit adds a `headers` parameter to the `FileTransferRequest`, and
the github fetchers now call it with
`Authorization: token <github-access-token>`.

This differs from the current convention in nix of passing the access
token as a query parameter, because query parameter tokens are
deprecated and will be disallowed in November 2020. Using them today
triggers a warning email.